requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
qdrant-client
langchain_google_genai
python-dotenv
langchain_community
